+++ packages/browser-app/src/components/podcast-responder/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PodcastResponder';
