+++ packages/browser-app/src/components/podcast-responder/types.ts
@@ -1,0 +1,14 @@
+export interface MediaSource {
+  title: string;
+  url: string;
+  author?: string;
+  notes?: string;
+}
+
+export interface Message {
+  id: string;
+  role: 'user' | 'assistant';
+  content: string;
+}
+
+export type Step = 'input' | 'chat' | 'preview';
